+++ src/helix/mod.rs
@@ -54,6 +54,7 @@ pub mod clips;
 #[cfg_attr(nightly, doc(cfg(feature = "eventsub")))]
 pub mod eventsub;
 pub mod games;
+pub mod goals;
 pub mod hypetrain;
 pub mod moderation;
 pub mod points;